removed fragments rendered useless after creating new review fragment
