+++ app/src/main/java/com/example/andresarango/aughunt/CreateChallengeActivity.java
@@ -5,8 +5,8 @@ import android.support.v7.app.AppCompatActivity;
 
 import com.example.andresarango.aughunt.challenge.ChallengePhoto;
 import com.example.andresarango.aughunt.challenge.ChallengePhotoCompleted;
-import com.example.andresarango.aughunt.challenge.challenge_review_fragments.CompareChallengesFragment;
 import com.example.andresarango.aughunt.challenge.challenge_review_fragments.CreatedChallengesFragment;
+import com.example.andresarango.aughunt.challenge.challenge_review_fragments.PopFragmentListener;
 import com.example.andresarango.aughunt.challenge.challenge_review_fragments.ReviewChallengesFragment;
 import com.example.andresarango.aughunt.challenge.challenges_adapters.created.CreatedChallengeListener;
 import com.example.andresarango.aughunt.challenge.challenges_adapters.review.CompletedChallengeListener;
@@ -15,13 +15,10 @@ import butterknife.ButterKnife;
 
 
 public class CreateChallengeActivity extends AppCompatActivity implements
-        CreatedChallengeListener,
-        CompletedChallengeListener {
+        CreatedChallengeListener, PopFragmentListener {
 
     private CreatedChallengesFragment mCreatedChallengesFragment;
     private ReviewChallengesFragment mReviewChallengesFragment;
-    private CompareChallengesFragment mCompareChallengesFragment;
-    private ChallengePhoto mSelectedChallenge;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -39,19 +36,13 @@ public class CreateChallengeActivity extends AppCompatActivity implements
 
     @Override
     public void onCreatedChallengeClicked(ChallengePhoto challenge) {
-        mSelectedChallenge = challenge;
         startReviewChallengeFragment(challenge);
-    }
-
-    @Override
-    public void onCompletedChallengeClicked(ChallengePhotoCompleted completedChallenge) {
-        startCompareChallengeFragment(completedChallenge, mSelectedChallenge);
     }
 
     private void startReviewChallengeFragment(ChallengePhoto challenge) {
         mReviewChallengesFragment = new ReviewChallengesFragment();
         mReviewChallengesFragment.setChallengeToReview(challenge);
-        mReviewChallengesFragment.setmListener(this);
+        mReviewChallengesFragment.setPopFragmentListener(this);
 
         getSupportFragmentManager()
                 .beginTransaction()
@@ -61,19 +52,15 @@ public class CreateChallengeActivity extends AppCompatActivity implements
     }
 
 
-    private void startCompareChallengeFragment(ChallengePhotoCompleted completedChallenge, ChallengePhoto challenge) {
-        mCompareChallengesFragment = new CompareChallengesFragment();
-        mCompareChallengesFragment.setCompletedChallenge(completedChallenge);
-        mCompareChallengesFragment.setCurrentChallenge(challenge);
-        getSupportFragmentManager()
-                .beginTransaction()
-                .replace(R.id.container_for_review, mCompareChallengesFragment)
-                .addToBackStack(null)
-                .commit();
-    }
 
     public void popFragmentFromBackStack() {
         getSupportFragmentManager().popBackStack();
     }
+
+    @Override
+    public void popFragment() {
+        popFragmentFromBackStack();
+    }
+
 
 }
