+++ app/src/main/java/com/example/andresarango/aughunt/ProfileActivity.java
@@ -19,10 +19,9 @@ import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.target.BitmapImageViewTarget;
 import com.example.andresarango.aughunt.challenge.ChallengePhoto;
 import com.example.andresarango.aughunt.challenge.ChallengePhotoCompleted;
-import com.example.andresarango.aughunt.challenge.challenge_review_fragments.CompareChallengesFragment;
 import com.example.andresarango.aughunt.challenge.challenge_review_fragments.CreatedChallengesFragment;
 import com.example.andresarango.aughunt.challenge.challenge_review_fragments.PopFragmentListener;
-import com.example.andresarango.aughunt.challenge.challenge_review_fragments.ReviewChallengesFragmentAndres;
+import com.example.andresarango.aughunt.challenge.challenge_review_fragments.ReviewChallengesFragment;
 import com.example.andresarango.aughunt.challenge.challenges_adapters.created.CreatedChallengeListener;
 import com.example.andresarango.aughunt.challenge.challenges_adapters.review.CompletedChallengeListener;
 import com.example.andresarango.aughunt.profile.ViewPagerAdapter;
@@ -41,20 +40,25 @@ import butterknife.ButterKnife;
  * Created by dannylui on 3/11/17.
  */
 
-public class ProfileActivity extends AppCompatActivity implements CreatedChallengeListener, CompletedChallengeListener,PopFragmentListener {
-    @BindView(R.id.tab_layout) TabLayout tablayout; // Import design in build.gradle
-    @BindView(R.id.viewpager) ViewPager pager;
-    @BindView(R.id.bottom_navigation) BottomNavigationView mBottomNav;
-    @BindView(R.id.iv_main_profile_pic) ImageView profilePicIv;
-    @BindView(R.id.tv_main_profile_points) TextView userPointsTv;
-    @BindView(R.id.tv_main_profile_total_created) TextView totalCreatedChallengesTv;
-    @BindView(R.id.tv_main_profile_total_submitted) TextView totalSubmittedChallengesTv;
+public class ProfileActivity extends AppCompatActivity implements CreatedChallengeListener, PopFragmentListener {
+    @BindView(R.id.tab_layout)
+    TabLayout tablayout; // Import design in build.gradle
+    @BindView(R.id.viewpager)
+    ViewPager pager;
+    @BindView(R.id.bottom_navigation)
+    BottomNavigationView mBottomNav;
+    @BindView(R.id.iv_main_profile_pic)
+    ImageView profilePicIv;
+    @BindView(R.id.tv_main_profile_points)
+    TextView userPointsTv;
+    @BindView(R.id.tv_main_profile_total_created)
+    TextView totalCreatedChallengesTv;
+    @BindView(R.id.tv_main_profile_total_submitted)
+    TextView totalSubmittedChallengesTv;
 
 
     private CreatedChallengesFragment mCreatedChallengesFragment;
-    private ReviewChallengesFragmentAndres mReviewChallengesFragment;
-    private CompareChallengesFragment mCompareChallengesFragment;
-    private ChallengePhoto mSelectedChallenge;
+    private ReviewChallengesFragment mReviewChallengesFragment;
 
     private FirebaseAuth auth = FirebaseAuth.getInstance();
     private DatabaseReference rootRef = FirebaseDatabase.getInstance().getReference();
@@ -74,20 +78,20 @@ public class ProfileActivity extends AppCompatActivity implements CreatedChallen
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
 
-                switch (item.getItemId()){
+                switch (item.getItemId()) {
 
                     case R.id.create_challenge:
-                        Intent createChallenge=new Intent(getApplicationContext(), ChallengeTemplateActivity.class);
+                        Intent createChallenge = new Intent(getApplicationContext(), ChallengeTemplateActivity.class);
                         startActivity(createChallenge);
 
                         break;
                     case R.id.homepage:
-                        Intent homePage=new Intent(getApplicationContext(), SearchChallengeActivity.class);
+                        Intent homePage = new Intent(getApplicationContext(), SearchChallengeActivity.class);
                         startActivity(homePage);
 
                         break;
                     case R.id.user_profile:
-                        Intent userProfile=new Intent(getApplicationContext(),ProfileActivity.class);
+                        Intent userProfile = new Intent(getApplicationContext(), ProfileActivity.class);
                         startActivity(userProfile);
                         break;
                 }
@@ -153,34 +157,17 @@ public class ProfileActivity extends AppCompatActivity implements CreatedChallen
 
     @Override
     public void onCreatedChallengeClicked(ChallengePhoto challenge) {
-        mSelectedChallenge = challenge;
         startReviewChallengeFragment(challenge);
     }
 
     private void startReviewChallengeFragment(ChallengePhoto challenge) {
-        mReviewChallengesFragment = new ReviewChallengesFragmentAndres();
+        mReviewChallengesFragment = new ReviewChallengesFragment();
         mReviewChallengesFragment.setChallengeToReview(challenge);
         mReviewChallengesFragment.setPopFragmentListener(this);
 
         getSupportFragmentManager()
                 .beginTransaction()
                 .replace(R.id.profile_activity, mReviewChallengesFragment)
-                .addToBackStack(null)
-                .commit();
-    }
-
-    @Override
-    public void onCompletedChallengeClicked(ChallengePhotoCompleted completedChallenge) {
-        startCompareChallengeFragment(completedChallenge, mSelectedChallenge);
-    }
-
-    private void startCompareChallengeFragment(ChallengePhotoCompleted completedChallenge, ChallengePhoto challenge) {
-        mCompareChallengesFragment = new CompareChallengesFragment();
-        mCompareChallengesFragment.setCompletedChallenge(completedChallenge);
-        mCompareChallengesFragment.setCurrentChallenge(challenge);
-        getSupportFragmentManager()
-                .beginTransaction()
-                .replace(R.id.profile_activity, mCompareChallengesFragment)
                 .addToBackStack(null)
                 .commit();
     }
